% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipalities.R
\name{check_year_}
\alias{check_year_}
\title{Internal function check_year_}
\usage{
check_year_(year, argument)
}
\arguments{
\item{year}{Year}

\item{argument}{Name of the argument being checked, used in the error message}
}
\description{
Stop unless `year` is a single year within the range covered by this
package. Without this check, a vector, NA or character year reaches
`if` or `dplyr::between` and fails with an opaque message.
}
