% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipalities.R
\name{check_years_}
\alias{check_years_}
\title{Internal function check_years_}
\usage{
check_years_(years)
}
\arguments{
\item{years}{Years}
}
\description{
Stop unless `years` is a non-empty vector of years within the range
covered by this package. The error names up to five offending years.
}
