% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counties.R
\docType{data}
\name{counties}
\alias{counties}
\title{NJ Counties}
\format{
A dataframe with 21 rows and 2 columns
\describe{
  \item{GEOID}{US Census Geographical Identifier}
  \item{County}{County name}
}
}
\usage{
counties
}
\description{
All 21 NJ counties and their GEOIDs.
}
\keyword{datasets}
