% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipalities.R
\name{get_geoid_cross_references}
\alias{get_geoid_cross_references}
\title{Get a cross-reference table for GEOIDs}
\usage{
get_geoid_cross_references(reference_year, years)
}
\arguments{
\item{reference_year}{Reference year}

\item{years}{Years}
}
\value{
Cross reference table
}
\description{
Given a reference year and a list of years, return a table of
cross references of the form `~year, ~GEOID_ref, ~GEOID`
}
\examples{
# Cross reference 2012 and 2013 GEOIDs against a 2010 reference,
# and sort by the reference GEOID.
get_geoid_cross_references(2010, 2012:2013) \%>\%
  dplyr::arrange(GEOID_ref)
}
