% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_municipalities.R
\name{get_municipalities_}
\alias{get_municipalities_}
\title{Internal function get_municipalities_}
\usage{
get_municipalities_(year = MAX_YEAR, geoid_reference_year = year)
}
\arguments{
\item{year}{Year}

\item{geoid_reference_year}{GEOID reference year}
}
\description{
Given a year and a reference year, return a list of all
municipalities matching the year, and include GEOID of the
reference year as a separate column, GEOID_ref.
}
