% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PINE_HILL_BORO_GEOID.R
\docType{data}
\name{PINE_HILL_BORO_GEOID}
\alias{PINE_HILL_BORO_GEOID}
\title{GEOID for Pine Hill borough}
\format{
An object of class \code{character} of length 1.
}
\usage{
PINE_HILL_BORO_GEOID
}
\description{
GEOID for Pine Hill borough
}
\keyword{datasets}
