% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRINCETON_BORO_GEOID.R
\docType{data}
\name{PRINCETON_BORO_GEOID}
\alias{PRINCETON_BORO_GEOID}
\title{GEOID for Princeton borough}
\format{
An object of class \code{character} of length 1.
}
\usage{
PRINCETON_BORO_GEOID
}
\description{
GEOID for Princeton borough
}
\keyword{datasets}
