% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PRINCETON_TWP_GEOID.R
\docType{data}
\name{PRINCETON_TWP_GEOID}
\alias{PRINCETON_TWP_GEOID}
\title{GEOID for Princeton township}
\format{
An object of class \code{character} of length 1.
}
\usage{
PRINCETON_TWP_GEOID
}
\description{
GEOID for Princeton township
}
\keyword{datasets}
